Proposal for new parser infrastructure
