+++ packages/langium/src/index.ts
@@ -12,7 +12,6 @@ export * from './generator/generator-node';
 export * from './generator/node-processor';
 export * from './generator/template-string';
 export * from './grammar/generated/ast';
-export * from './grammar/generated/grammar-access';
 export * from './grammar/generated/module';
 export * from './grammar/generated/parser';
 export * from './grammar/grammar-access';
@@ -25,7 +24,7 @@ export * from './lsp/completion/rule-interpreter';
 export * from './lsp/language-server';
 export * from './validation/document-validator';
 export * from './validation/validation-registry';
-export * from './parser/langium-parser';
+export * from './parser/langium-base-parser';
 export * from './references/linker';
 export * from './references/naming';
 export * from './references/scope';
